Add basic dependencies and installation instruction
Poetry does good job for dependencies management. The auto-generated
`poetry.lock` is also version controlled to have frozen version of
packages used for the original development and testing.

This project was created using python 3.8 currently installed on my
system, so it was tested only with this version.

Because of the I/O bound nature of the project, asyncio could be best
option for it. This is the reason for aiokafka, httpx and asyncpg being
chosen as basic libraries.

tidypy was chosen as a basic code quality analysis framework (e.g.
instead of prospector) because I wanted to try it and it has integration
with pytest and pyproject.toml. Some invaluable or not very practical
checks were disabled (unfortunately I didn't document names).

Signed-off-by: Denis Zalevskiy <denis@visfun.org>